Working on removing auditing direct integration with persistence and adding the functionality to the auditing module instead. Other modules will need to have the changes to persistence module applied and the auditing module integrated where appropriate.
